server/pyproject.toml
[project]
name = "eyenav"
version = "0.1.0"
description = "Navigate your device effortlessly with your eyes"
authors = [
    {name = "Manasseh",email = "believemanasseh@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "transformers (==4.49.0)",
    "pillow (==11.1.0)",
    "torch (==2.6.0)",
    "mediapipe (==0.10.21)",
    "opencv-python (==4.11.0.86)",
    "uvicorn (==0.34.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
